requirement.txt
 pandas==2.2.3
 plotly_express==0.4.1
 streamlit==1.37.1